Initial conversion of rtld object permutation randomizing to use fisher-yates
This *should* mean that the permutation of objects we wind up
with should be 'chosen' uniformly from the set of all possible
permutations. Before, the object shuffling algorithm wasn't
guaranteed to uniformly shuffle the objects, meaning some
permutations were likelier than others. Maybe not a big weakness,
but certainly something fixable.

This includes a uniform_random function which, given N, returns
a number chosne uniformly from [0,N). This is critical to fisher-yates,
and should be very closely inspected for bugs.